Fix stats ping
The download-complete/installer-run stats pings weren't being sent
when the stub installer was run in admin mode. This was due to us
storing the app guid in the registry, as it is written to a different
location when running as an admin vs a regular user. Fixed this by
fetching app guid from tag instead. In addition, we append the
referral code to the tag sent to the installer in order to avoid the
same problem with the registry.