Remove stub silent installer
Before this commit, BUILD.gn / build_omaha.py generated a "stub silent
installer". This was a 1 MB online installer that fetched the latest
version of (in our case) Brave from the update server. The problem was
that this installer wasn't truly silent because it opened a browser
window. It was decided that the stub installer is not needed because
there is also a standalone (i.e. offline) silent installer that does
not open a browser window.